Reactor library wrapping EventBase
Summary:
Wrap EventBase to be used as the core of a Reactor pattern.

We couldn't write something in pure Lua as we need a way to schedule a callback
from C++ code running in another thread. Wrapping EventBase is useful as we
can integrate with Thrift async callbacks further down the road.

Test Plan: test added

Reviewed By: tulloch@fb.com

Subscribers: lesha, jhj, fjargsto

FB internal diff: D1488378