Async random number generation on the host
Summary:
Generate random numbers in different threads. All C++11 distributions are
supported. The numbers are not reproducible, as the per-thread generators
have different seeds and there's no way to control the interleaving.

This is similar in spirit and implementation to the async batch loader.

This gives about 9-10 ms / minibatch in imagenet, and it's also useful on
its own.

Test Plan: by hand, imagenet

Reviewed By: tulloch@fb.com

Subscribers: tulloch, ntv, fjargsto, jhj, lesha, spiantino

FB internal diff: D1502318