`defaultdict` in Lua, similar to Python's `collections.defaultdict`
Summary:
This is a useful primitive in general IMO - there have been a
bunch of cases where code has been reaching for something like this.

I'm happy to put it elsewhere - it's currently used in a few files in
an upcoming diff but I can just make a utility file for those.

Test Plan: I should write unit tests for this actually.  I'll do that.

Reviewed By: tudorb@fb.com

Subscribers: tudor, trunkagent, tulloch, pamelavagata, ntv, jhj

FB internal diff: D1619684

Tasks: 5336397