word2torch, and some miscellaneous library fixes.
Summary:This is a crude knock-off of word2vec rendered in torch. Salient
differences are that it uses a fixed window size, only does the skipgram
task, and uses a wsabie-style loss instead of an HSM.

The immediate goal is to use this as a test workload for distributed
embedding training.

Working with strdict and atomicvec in more depth exposed a few
opportunities in these modules and their corresponding support code.

Reviewed By: soumith

Differential Revision: D1875415

fb-gh-sync-id: rFBCODE58f74835bda04dac38324dc3672aacdbb31443af
shipit-source-id: rFBCODE58f74835bda04dac38324dc3672aacdbb31443af