fb.eval_none for roundtripping py.None
Summary:lvdmaaten had a case with a datastructure in Python like:

```
{"a": None, "b":2}
```

that gets `py.eval`'ed into

```
{b=2}
```

and we lose the information that `a=py.None` was set, since we can't have `nil` table values in Lua.

While one workaround is to return the key set and the dict to Lua,
another would be to have an eval mode that maps Python `None` into Lua `py.None`.

This diff does that, so we can have code like:

```
    assertEquals(
        {a=2,b={c=py.None}},
        py.eval_none('a', {a=py.eval_none('{"a": 2, "b": {"c": None}}')}))
```

Thoughts?

Reviewed By: tudor

Differential Revision: D2229391

fb-gh-sync-id: c85800d8b2ba5c2c5feed7d2a9f4c0206bf41b84
shipit-source-id: c85800d8b2ba5c2c5feed7d2a9f4c0206bf41b84