Make Reactor work with coroutines
Summary:No longer cache a lua_State*, but be careful to use the passed-in lua_State* at
all times. Note that, in functions scheduled using the Executor interface,
there's no easy way to determine the correct lua_State* (that the Reactor is
looping in), so keep a thread-local stack and make it accessible.

Reviewed By: nicolasvasilache

Differential Revision: D2604214

fb-gh-sync-id: 1706664a4a96c34434f6467b8679a5e9fc2e5d1a
shipit-source-id: 1706664a4a96c34434f6467b8679a5e9fc2e5d1a