Fix circular loading dependency between fb.trepl and fb.debugger
Summary:
Lazy load fb.trepl from fb.debugger to avoid circular loading error.

There's been a circular loading dependency error between fb.trepl and fb.debugger for ~forever, but it was hidden in a pcall. This was silently ignored because the code flow used to be trepl -> debugger [pcall] -> (trepl!), which triggered the error within the pcall in debugger. But D3511915 changed the order to base -> debugger [pcall] -> trepl -> (debugger!) which triggers the error first in trepl. This error is silently ignored in the pcall, but future requires to fb.trepl fail because the error is cached (unlike in the first case).

Reviewed By: tudor

Differential Revision: D3529785

fbshipit-source-id: 17429e3f3f979ff26477d9aa504c1d275198aa5c