Save history after every line
Summary: cuth/th isn't currently saving history. The GC hook on EditLine.hist_event is never called, so history is never saved to a file. This changes add_history to immediately save the history to the ~/.luahistory file (which is the same behavior as trepl). This is also nice because it means history isn't lost if the repl crashes.

Differential Revision: D3664334

fbshipit-source-id: bcf832a7a3b9af45c47766474f5ba8e0a68136ea